refactored classes into their own files
