+++ lib/code_generator.rb
@@ -1,0 +1,7 @@
+class CodeGenerator
+
+	def generate (colors)
+		4.times.map { |c| colors.sample }
+	end
+
+end
